create new class Interface and moving signaling process functions from class UDS into it. Create application entry file - main.cpp.
