+++ car/src/interface/src/interface.cpp
@@ -1,35 +1,100 @@
+#include "interface.h"
 
-#include <cstdlib>
-#include <string>
-#include <thread>
-#include "ros/ros.h"
-
-#include "uds.h"
-#include "roscar_common/rcmp.h"
+#include "rapidjson/writer.h"
 
 using namespace std;
-using namespace roscar::car::interface;
+using namespace rapidjson;
 using namespace roscar::car::roscar_common;
 
-void udsThreadFunc(const char *uri)
+namespace roscar
 {
-    ROS_INFO("start Unix Domain Socket on %s", uri);
-}
+namespace car
+{
+namespace interface
+{
 
-int main(int argc, char **argv)
+bool Interface::onSignaling(UDS::SESSION_t &sess, rapidjson::Document &sig)
 {
-    ros::init(argc, argv, "interface");
-    if (argc != 1)
+    const char *cmd = sig[RCMP::FIELD_CMD].GetString();
+
+    if (strcasecmp(cmd, RCMP::SIG_LOGIN_RESP))
     {
-        ROS_INFO("usage: interface");
-        return 1;
+        // In current version of code, skip signaling SIG_LOGIN_RESP.
+        ROS_DEBUG("Debug: In current version of code, skip signaling SIG_LOGIN_RESP.");
+        return true;
+    }
+    else if (strcasecmp(cmd, RCMP::SIG_PING))
+    {
+        return onSigPing(sess, sig);
+    }
+    else if (strcasecmp(cmd, RCMP::SIG_PONG))
+    {
+        return onSigPong(sess, sig);
+    }
+    else if (strcasecmp(cmd, RCMP::SIG_LOGIN))
+    {
+        // these signaling should not be received
+        ROS_ERROR("Err: wrong signaling(cmd: [%s]", cmd);
+        return false;
     }
 
-    UDS uds;
-    RCMP rcmp;
-
-    thread udsThread = thread(&UDS::threadFunc, &uds);
-    udsThread.join();
-
-    return 0;
+    // unsupported singaling
+    ROS_ERROR("Err: unsupported signaling(cmd: [%s]", cmd);
+    return false;
 }
+
+bool Interface::onSigPing(UDS::SESSION_t &sess, rapidjson::Document &sig)
+{
+    mRcmp.convertToPong(sig);
+
+    StringBuffer buffer;
+    Writer<StringBuffer> writer(buffer);
+    sig.Accept(writer);
+    const char *payload = buffer.GetString();
+
+    // send signaling via append it to send buffer
+    return sendSignaling(sess, buffer);
+}
+
+bool Interface::onSigPong(UDS::SESSION_t &sess, rapidjson::Document &sig)
+{
+    // In current version of code, we just skip signaling PONG.
+    ROS_DEBUG("Debug: In current version of code, skip signaling PONG.");
+    return true;
+}
+
+bool Interface::sendSignaling(UDS::SESSION_t &sess, StringBuffer &buffer)
+{
+    auto len = buffer.GetSize();
+
+    // check empty send buffer size
+    if (RCMP::RCMP_MAXPAYLOAD - sess.sendBufEnd < len)
+    {
+        if (sess.sendBufPos == 0)
+        {
+            // send buffer full
+            ROS_DEBUG("Debug: send buffer fulll(soc: [%d]", sess.soc);
+            return false;
+        }
+
+        // defrag buffer
+        int bufLen = sess.sendBufEnd - sess.sendBufPos;
+        memmove(sess.sendBuf, sess.sendBuf + sess.sendBufPos, bufLen);
+        sess.sendBufPos = 0;
+        sess.sendBufEnd = bufLen;
+
+        if (RCMP::RCMP_MAXPAYLOAD - sess.sendBufEnd < len)
+        {
+            // buffer still full after defrag send buffer
+            ROS_DEBUG("Debug: send buffer fulll(soc: [%d]", sess.soc);
+            return false;
+        }
+    }
+
+    memcpy(sess.sendBuf + sess.sendBufEnd, buffer.GetString(), len);
+    sess.sendBufEnd += len;
+}
+
+} // namespace interface
+} // namespace car
+} // namespace roscar
